requirements.txt
flask==3.0.3
yfinance==1.2.0
python-dotenv==1.0.1
pytest==8.2.2
